Brother and seeporn fixes
Fixed a second issue in seeporn that caused it to crash when it got to a
certain point.
Did minor cleanup on action names in seeporn to make them a bit clearer
as to what they are doing.
Moved rand call in seeporn into if block.
Some cleanup in brother to eliminate unnecessary else condition, using
exit instead to cut down on nesting. Moved some commands around to
further simplify code, also moved a rand call into if condition.